Significant changes to client to help integrate better with TUI: WIP
In RSyncer stdout parsing I think it needs to check if line starts with >f rather than <f
