+++ src/murfey/client/watchdir.py
@@ -7,7 +7,7 @@ from pathlib import Path
 from typing import NamedTuple, Optional
 
 import murfey.util
-from murfey.client.tui import StatusBar
+from murfey.client.tui.status_bar import StatusBar
 
 log = logging.getLogger("murfey.client.watchdir")
 
@@ -36,45 +36,56 @@ class DirWatcher(murfey.util.Observer):
         return f"<DirWatcher ({self._basepath})>"
 
     def scan(self):
-        t_start = time.perf_counter()
-        filelist = self._scan_directory()
-        t_scan = time.perf_counter() - t_start
-        log.info(f"Scan of {self._basepath} completed in {t_scan:.1f} seconds")
-        scan_completion = time.time()
+        try:
+            t_start = time.perf_counter()
+            filelist = self._scan_directory()
+            t_scan = time.perf_counter() - t_start
+            log.info(f"Scan of {self._basepath} completed in {t_scan:.1f} seconds")
+            scan_completion = time.time()
 
-        for entry, entry_info in filelist.items():
-            if entry_info != self._lastscan.get(entry):
-                self._file_candidates[entry] = entry_info._replace(
-                    settling_time=scan_completion
-                )
+            for entry, entry_info in filelist.items():
+                if entry_info != self._lastscan.get(entry):
+                    self._file_candidates[entry] = entry_info._replace(
+                        settling_time=scan_completion
+                    )
 
-        for x in sorted(self._file_candidates):
-            if x not in filelist:
-                log.info(f"Previously seen file {x!r} has disappeared")
-                del self._file_candidates[x]
-                continue
-
-            if (
-                self._file_candidates[x].settling_time + self.settling_time
-                < time.time()
-            ):
-                file_stat = os.stat(x)
-                if (
-                    file_stat.st_size == self._file_candidates[x].size
-                    and file_stat.st_mtime <= self._file_candidates[x].modification_time
-                    and file_stat.st_ctime <= self._file_candidates[x].modification_time
-                ):
-                    log.debug(f"File {x!r} is ready to be transferred")
-                    if self._statusbar:
-                        self._statusbar.transferred[1] += 1
-                    self.notify(Path(x))
+            for x in sorted(self._file_candidates):
+                if x not in filelist:
+                    log.info(f"Previously seen file {x!r} has disappeared")
                     del self._file_candidates[x]
                     continue
 
-            if x not in self._lastscan:
-                log.debug(f"Found file {x!r} for future transfer")
+                if (
+                    self._file_candidates[x].settling_time + self.settling_time
+                    < time.time()
+                ):
+                    try:
+                        file_stat = os.stat(x)
+                        if (
+                            file_stat.st_size == self._file_candidates[x].size
+                            and file_stat.st_mtime
+                            <= self._file_candidates[x].modification_time
+                            and file_stat.st_ctime
+                            <= self._file_candidates[x].modification_time
+                        ):
+                            log.debug(
+                                f"File {Path(x).name!r} is ready to be transferred"
+                            )
+                            if self._statusbar:
+                                self._statusbar.transferred[1] += 1
+                            self.notify(Path(x))
+                            del self._file_candidates[x]
+                            continue
+                    except Exception as e:
+                        log.error(f"Exception encountered: {e}", exc_info=True)
+                        return
 
-        self._lastscan = filelist
+                if x not in self._lastscan:
+                    log.debug(f"Found file {Path(x).name!r} for future transfer")
+
+            self._lastscan = filelist
+        except Exception as e:
+            log.error(f"Exception encountered: {e}")
 
     def _scan_directory(self, path: str = "") -> dict[str, _FileInfo]:
         result: dict[str, _FileInfo] = {}
@@ -100,7 +111,7 @@ class DirWatcher(murfey.util.Observer):
                     # between the scandir and the stat call.
                     # In this case we can just ignore the file.
                     continue
-                result[entry_name] = _FileInfo(
+                result[str(Path(self._basepath) / path / entry_name)] = _FileInfo(
                     size=file_stat.st_size,
                     modification_time=max(file_stat.st_mtime, file_stat.st_ctime),
                 )
